requirements.txt
pytest
pytest-html
python-dateutil


# Agrega otras dependencias según sea necesario

mutmut